refactor(plt-2268): change file naming convention
